Refactored callbacks config -> directly return list instead of dict, move indiv callbacks dataclass to separate file
